refactor(analytics): migrate search history and application metrics to Elasticsearch mappings, removing deprecated entities and services
