+++ src/analytics/schema/application-metrics.schema.ts
@@ -1,0 +1,10 @@
+
+export enum ApplicationProgressStep {
+    APPLICATION_START = 'application/start',
+    PROFILE_SELF = 'profile/self',
+    PROFILE_CONTACT = 'profile/contact',
+    PROFILE_EDUCATION = 'profile/education',
+    PROFILE_DOCS = 'profile/docs',
+    APPLICATION_PROGRAM_SELECTION = 'application/program_selection',
+    APPLICATION_COMPLETE = 'application/complete',
+}
