+++ src/elasticsearch/mappings/base.mapping.ts
@@ -1,7 +1,7 @@
 import { EsEntity, EsField } from 'es-mapping-ts';
 
 @EsEntity()
-export class BaseEntity {
+export class BaseMappingEntity {
   @EsField({
     type: 'date'
   })
@@ -11,4 +11,9 @@ export class BaseEntity {
     type: 'date'
   })
   updated_at: Date;
+
+  @EsField({
+    type: 'date'
+  })
+  timestamp: Date;
 }
