+++ src/elasticsearch/mappings/es-indices.enum.ts
@@ -1,0 +1,7 @@
+/**
+ * Default Elasticsearch indices names
+ */
+export enum ES_INDICES {
+    SEARCH_HISTORY = 'search_history',
+    APPLICATION_METRICS = 'application_metrics',
+}
