+++ src/programs/entities/program.entity.ts
@@ -1,10 +1,10 @@
 import { EsEntity, EsField } from 'es-mapping-ts';
-import { BaseEntity } from '../../common/entities/base.entity';
+import { BaseMappingEntity } from '../../elasticsearch/mappings/base.mapping';
 
 @EsEntity({
   index: 'programs'
 })
-export class ProgramEntity extends BaseEntity {
+export class ProgramEntity extends BaseMappingEntity {
   @EsField({
     type: 'text',
     analyzer: 'english',
