+++ src/universities/universities.module.ts
@@ -1,12 +1,11 @@
 import { Module } from '@nestjs/common';
 import { MongooseModule } from '@nestjs/mongoose';
-import { UniversitiesService } from './universities.service';
-import { UniversitiesController } from './universities.controller';
-import { Program, ProgramSchema } from '../programs/schemas/program.schema';
-import { Campus, CampusSchema } from '../campuses/schemas/campus.schema';
 import { AnalyticsModule } from '../analytics/analytics.module';
+import { Campus, CampusSchema } from '../campuses/schemas/campus.schema';
 import { ElasticsearchModule } from '../elasticsearch/elasticsearch.module';
-import { MappingRegistryService } from 'src/elasticsearch/services/mapping-registry.service';
+import { Program, ProgramSchema } from '../programs/schemas/program.schema';
+import { UniversitiesController } from './universities.controller';
+import { UniversitiesService } from './universities.service';
 import { UniversityModelsModule } from './university-models.module';
 
 @Module({
@@ -20,7 +19,7 @@ import { UniversityModelsModule } from './university-models.module';
     ElasticsearchModule,
   ],
   controllers: [UniversitiesController],
-  providers: [UniversitiesService, MappingRegistryService],
+  providers: [UniversitiesService],
   exports: [UniversitiesService],
 })
 export class UniversitiesModule {}
